Add newer headling level toolbar from WP core
